+++ src/components/atom_word_card_delete_button/index.tsx
@@ -2,9 +2,9 @@ import StyledIconButtonAtom from '@/atoms/StyledIconButton'
 import { FC } from 'react'
 import DeleteWordIcon from '@mui/icons-material/Delete'
 import { useRecoilState } from 'recoil'
-import { wordsFamily } from '@/recoils/state_atoms/words.state'
 import { deleteWordByIdApi } from '@/api/words/delete-words.api'
 import { WordData } from '@/api/words/words.interface'
+import { wordsFamily } from '@/recoil/words.state'
 interface Props {
   wordId: string
 }
