+++ src/components/atom_word_card_undo_delete_button/index.tsx
@@ -1,7 +1,7 @@
 import { postWordApi } from '@/api/words/post-word.api'
 import { WordData } from '@/api/words/words.interface'
 import StyledTextButtonAtom from '@/atoms/StyledTextButton'
-import { wordsFamily } from '@/recoils/state_atoms/words.state'
+import { wordsFamily } from '@/recoil/words.state'
 import { FC } from 'react'
 import { useRecoilState } from 'recoil'
