+++ src/components/atom_word_cards_frame_refresh_button/index.tsx
@@ -1,8 +1,8 @@
 import StyledIconButtonAtom from '@/atoms/StyledIconButton'
-import { wordIdsState } from '@/recoils/state_atoms/words.state'
 import { FC } from 'react'
 import { useResetRecoilState } from 'recoil'
 import RefreshIcon from '@mui/icons-material/Refresh'
+import { wordIdsState } from '@/recoil/words.state'
 
 const WordCardsFrameRefreshButton: FC = () => {
   const resetWordIds = useResetRecoilState(wordIdsState)
