+++ src/components/organism_word_cards_frame/index.tsx
@@ -3,10 +3,10 @@ import WordCard from '../molecule_word_card'
 import { Stack, Box } from '@mui/material'
 import NewWordBox from '../molecule_new_word_box'
 import { useRecoilValue } from 'recoil'
-import { searchInputState } from '@/recoils/state_atoms/searchInput.state'
+import { searchInputState } from '@/recoil/searchInput.state'
 import WordCardsFrameSearchNotFound from './index.search_not_found'
-import { wordIdsState } from '@/recoils/state_atoms/words.state'
 import WordCardsFrameRefreshButton from '../atom_word_cards_frame_refresh_button'
+import { wordIdsState } from '@/recoil/words.state'
 
 const WordCardsFrame: FC = () => {
   const searchInput = useRecoilValue(searchInputState)
