+++ src/recoil/index.guard.ts
@@ -1,0 +1,8 @@
+import { DefaultValue } from 'recoil'
+
+export const guardRecoilDefaultValue = (
+  candidate: unknown,
+): candidate is DefaultValue => {
+  if (candidate instanceof DefaultValue) return true
+  return false
+}
