+++ src/recoil/searchInput.state.ts
@@ -1,5 +1,5 @@
 import { atom, RecoilState } from 'recoil'
-import { RecoilKey } from '../keys.recoil'
+import { RecoilKey } from './index.keys'
 
 export const searchInputState: RecoilState<string> = atom({
   key: RecoilKey.SearchInput,
